Adds hook to catch tags that don't match project version
Adds a pre-push hook to catch when tags are being pushed when they don't
match the project version. The hooks live in `git_hooks`, which should
be linked as `.git/hooks`. ON execution, the run

`cargo run -p hooks_executable -- $THE_HOOK_NAME`

and the executable switches logic on the hook name. Adding hooks is
a matter of implementing them in hook_executable and creating the
sylink with the appropriate hook name in `git_hooks/`